Impedindo a exclusão de um bem associado a um inventário.

+++ src/main/java/br/edu/ifsp/addthenewsoul/domain/usecases/asset/RemoveAssetUseCase.java
@@ -1,6 +1,7 @@
 package br.edu.ifsp.addthenewsoul.domain.usecases.asset;
 
 import br.edu.ifsp.addthenewsoul.domain.entities.asset.Asset;
+import br.edu.ifsp.addthenewsoul.domain.usecases.utils.exceptions.AssetBelongsToOpenedInventoryException;
 import br.edu.ifsp.addthenewsoul.domain.usecases.utils.exceptions.EntityAlreadyExistsException;
 
 public class RemoveAssetUseCase {
@@ -17,6 +18,14 @@ public class RemoveAssetUseCase {
         Integer id = asset.getId();
         if (assetDAO.findById(id).isEmpty())
             throw new EntityAlreadyExistsException("This resource ID does not exist");
+
+        boolean assetBelongsToInventory = assetDAO.findByIdWithInventoryAsset(id).get()
+                                            .getInventoryAsset() != null;
+        if (assetBelongsToInventory) {
+            throw new AssetBelongsToOpenedInventoryException("Não é possível remover um bem associado " +
+                                                              "a um inventário em andamento");
+        }
+
         assetDAO.delete(asset.getId());
     }
 }
